ELT_Pipeline_HG_Insight/requirements.txt
pandas
sqlalchemy
python-dotenv
pytest
faker
